actualizacion
se modidifcaron cosas y se agregaron otras falta lo de que  obtenr la
conver  no creo que este aca

+++ servidor TCP/src/servidorTCP.java
@@ -4,7 +4,7 @@ import java.util.StringTokenizer;
 
 public class servidorTCP implements Runnable {
 	
-	static final int puerto  = 8086;
+	static final int puerto  = 9046;
 	Socket con;
 	String protocol_set[] ={"hi_back","dispach_message","save_message"};
 	
@@ -22,17 +22,26 @@ public class servidorTCP implements Runnable {
 		DataOutputStream outClient =new DataOutputStream(this.con.getOutputStream());
 		String fichero ="src/"+ip_contacto+".txt";
 		File archivo = new File (fichero);
-		FileReader fr = new FileReader (archivo);
-		BufferedReader br = new BufferedReader(fr);
-		String linea;
+		if (archivo.exists()){
+			
+			FileReader fr = new FileReader (archivo);
+			BufferedReader br = new BufferedReader(fr);
+			String linea;
 		
-		while((linea=br.readLine())!=null){
-			 String mensaje = protocol_set[1]+"¬¬"+linea+"\n";
-	         outClient.writeBytes(mensaje);
-	         outClient.flush();
+			while((linea=br.readLine())!=null){
+				String mensaje = protocol_set[1]+"¬¬"+linea+"\n";
+				outClient.writeBytes(mensaje);
+				outClient.flush();
+			}
+			outClient.writeBytes("end"+"¬¬"+"chau"+"\n");
+			outClient.flush();
+			archivo.delete();
+		}	
+		else{
+			outClient.writeBytes("end"+"¬¬"+"chau"+"\n");
+			outClient.flush();
+			
 		}
-		outClient.writeBytes("end"+"\n");
-		archivo.delete();
 	}
 	
 	public void save_message(String IP,String message) throws IOException{
@@ -41,19 +50,11 @@ public class servidorTCP implements Runnable {
 		FileWriter escri = new FileWriter("src/" + fichero,true);
 		escri.write(IP+":"+message+"\r\n");
 		escri.close();
-		
-		//File chat = new File(IP+puerto+ip_contacto,port_contacto+".txt");
-		//FileWriter escribir = new FileWriter(chat, true);
-		//BufferedWriter escritor2 =new BufferedWriter(escribir);
-		//PrintWriter escritor3 = new PrintWriter(escritor2);
-		//escritor3.append(IP+":"+message+"\r\n");
-		//escritor3.close();
-		//escritor2.close();
-		
 		outClient.writeBytes(protocol_set[2]+"¬¬"+"mensaje recivido"+'\n');
 		outClient.flush();
 		
 	}
+	
 	public static void main(String argv[]) throws Exception{ 	 	
 			
 			ServerSocket socketservidor = new ServerSocket(puerto);
@@ -76,6 +77,9 @@ public class servidorTCP implements Runnable {
 				BufferedReader inFromClient = new BufferedReader(new InputStreamReader(con.getInputStream()));
 				clientSentence = inFromClient.readLine(); 
 				while(clientSentence != null){
+					
+					System.out.println(clientSentence);
+					
 					StringTokenizer tokens = new StringTokenizer(clientSentence,"¬¬");
 					String Protocolo =tokens.nextToken();
 					switch(Protocolo){
@@ -83,22 +87,30 @@ public class servidorTCP implements Runnable {
 						hi_back();
 						clientSentence = inFromClient.readLine();
 						System.out.println(clientSentence);
+						
 						break;
 					case("ask_message"):
-						String ip_salida =tokens.nextToken();
-						String puerto_salida=tokens.nextToken();
-						String ip_llegada=tokens.nextToken();
-						String puerto_llegada=tokens.nextToken();
-						String message=tokens.nextToken();
-						dispach_message(ip_salida);
+						
+						String ip_contacto=tokens.nextToken();
+						
+						//String message=tokens.nextToken();
+						System.out.println(ip_contacto);
+						dispach_message(ip_contacto);
+						 clientSentence = null;
+						System.out.println("guardado:"+clientSentence);
+		                   
+						
+						
+						
 						break;
 					case("send_message"):
+						
 						String ip_salida2 =tokens.nextToken();
-						String puerto_salida2=tokens.nextToken();
-						String ip_llegada2=tokens.nextToken();
-						String puerto_llegada2=tokens.nextToken();
 						String message2=tokens.nextToken();
 						save_message(ip_salida2,message2);
+						clientSentence = "rady";
+						System.out.println("ok"+clientSentence);
+						//System.out.println("se cae en el server");
 						break;
 						
 					}
